% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothExport.R
\name{DocExport}
\alias{DocExport}
\title{Export objects into a Word document}
\usage{
DocExport(filename, tables = NULL, tables_title = NULL,
  figures = NULL, figures_title = NULL, base_docx_filename = NULL)
}
\arguments{
\item{filename}{the name of the target Word document to create}

\item{tables}{a named list of flextable objects}

\item{tables_title}{a named list of table titles. The names should correspond to those of \emph{tables} parameter}

\item{figures}{a named list of \code{ggplot()} objects}

\item{figures_title}{a named list of figure titles. The names should correspond to those of \emph{figures} parameter}

\item{base_docx_filename}{the name of another Word file to inherit data from}
}
\value{
The message in console in case of successful Word document creation
}
\description{
Export flextable objects or ggplot2 figures into a Word document
}
\examples{
#export simple flextable objects
t_flx_obj <- CreateFLX(list(data = mtcars[1:8,]))
t_flname <- paste0(getwd(), "/test_table.docx")
DocExport(filename = t_flname,
          tables = list(table1 = t_flx_obj),
          tables_title = list(table1 = "Test table"))
t_flx_obj_sec <- CreateFLX(list(data = mtcars[9:16,]))
t_flname_sec <- paste0(getwd(), "/test_table_second.docx")
DocExport(filename = t_flname_sec,
          tables = list(table2 = t_flx_obj_sec),
          tables_title = list(table2 = "Test table second"),
          base_docx_filename = t_flname)
}
