% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothExport.R
\name{DocExportStyles}
\alias{DocExportStyles}
\title{Default styles for objects exporting}
\usage{
DocExportStyles()
}
\value{
set of styles
}
\description{
Store default styles settings for flextable or figures exporting into Word document
}
