% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothExport.R
\name{MergeCells}
\alias{MergeCells}
\title{Merges cells, rows and columns of a flextable object}
\usage{
MergeCells(df_flex, is_footer, columns, rows, cells)
}
\arguments{
\item{df_flex}{a flextable object}

\item{is_footer}{whether there is a footer in a flextable object}

\item{columns}{indexes of the columns to merge}

\item{rows}{indexes of the rows to merge}

\item{cells}{list of rectangulars as subdataframe to merge}
}
\value{
The flextable object with merged cells
}
\description{
Merges cells, rows and columns of a flextable object
}
