% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothExport.R
\name{SetBorders}
\alias{SetBorders}
\title{Set cell borders of a flextable object}
\usage{
SetBorders(df_flex)
}
\arguments{
\item{df_flex}{a flextable object}
}
\value{
The flextable object with borders
}
\description{
Set cell borders for header, footer and body parts of the flextable object
}
