% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothExport.R
\name{SetStyle}
\alias{SetStyle}
\title{Set styles of flextable cells}
\usage{
SetStyle(df_flex, font, rows_as_footnote_title, rows_as_inner_headers,
  rows_section_init, columns_left_align, rows_bold_upper_bordered,
  cells_color, rows_p_value_color)
}
\arguments{
\item{df_flex}{a flextable object}

\item{font}{font size}

\item{rows_as_footnote_title}{footer rows corresponding titles (Note, Designation, etc.)}

\item{rows_as_inner_headers}{body rows with inner headers (names of features, characteristics, etc)}

\item{rows_section_init}{body rows representing new section beginning (Normality test, etc)}

\item{columns_left_align}{columns to align left}

\item{rows_bold_upper_bordered}{rows with bold upper border}

\item{cells_color}{rows coloured orange (i.e. for a dropouts table)}

\item{rows_p_value_color}{body rows consisting of p_value for colouring}
}
\value{
The styled flextable object
}
\description{
Set styles of flextable cells
}
\details{
For more detailed parameters description see \code{CreateFLX()}.
}
