% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothExport.R
\name{SetWidthCells}
\alias{SetWidthCells}
\title{Set cell widths of a flextable object}
\usage{
SetWidthCells(df_flex, widths)
}
\arguments{
\item{df_flex}{a flextable object}

\item{widths}{a vector of widths in cm}
}
\value{
The flextable object with the fitted columns
}
\description{
Set cell widths of a flextable object
}
