% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothExport.R
\docType{package}
\name{SmoothExport}
\alias{SmoothExport}
\alias{SmoothExport-package}
\title{SmoothExport: package to convert R dataframes into flextable objects with capability of further export into Word files}
\description{
SmoothExport: package to convert R dataframes into flextable objects with capability of further export into Word files
}
